requirements.txt
# Voice Assistant Dependencies - API-Based (Simple & Easy)
# Compatible with Python 3.10.6

# Voice Recognition & Text-to-Speech
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13

# Web Automation
pywhatkit==5.4

# API-Based LLM (Choose one - easiest option)
openai==1.3.0
anthropic==0.7.0

# Optional: Groq (Fast & Free API)
groq==0.4.2

# Vector Storage for RAG Memory (Optional)
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Utilities
python-dotenv==1.0.0
pytz==2023.3
psutil==5.9.5
requests==2.31.0